+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -1191,15 +1191,21 @@ internal constructor(
      * @param reaction The [Reaction] to send.
      * @param enforceUnique Flag to determine whether the reaction should replace other ones added by the current user.
      * @param cid The full channel id, i.e. "messaging:123" to which the message with reaction belongs.
+     * @param skipPush If set to "true", skips sending push notification when reacting to a message.
      *
      * @return Executable async [Call] responsible for sending the reaction.
      */
     @CheckResult
     @JvmOverloads
-    public fun sendReaction(reaction: Reaction, enforceUnique: Boolean, cid: String? = null): Call<Reaction> {
+    public fun sendReaction(
+        reaction: Reaction,
+        enforceUnique: Boolean,
+        cid: String? = null,
+        skipPush: Boolean = false,
+    ): Call<Reaction> {
         val currentUser = getCurrentUser()
         val finalReaction = reaction.copy(createdLocallyAt = now())
-        return api.sendReaction(finalReaction, enforceUnique)
+        return api.sendReaction(finalReaction, enforceUnique, skipPush)
             .retry(scope = userScope, retryPolicy = retryPolicy)
             .doOnStart(userScope) {
                 logger.v { "[sendReaction] #doOnStart; reaction: ${reaction.type}, messageId: ${reaction.messageId}" }
@@ -1208,6 +1214,7 @@ internal constructor(
                         cid = cid,
                         reaction = finalReaction,
                         enforceUnique = enforceUnique,
+                        skipPush = skipPush,
                         currentUser = currentUser!!,
                     )
                 }
@@ -1224,7 +1231,7 @@ internal constructor(
                     )
                 }
             }
-            .onReactionError(errorHandlers, reaction, enforceUnique, currentUser!!)
+            .onReactionError(errorHandlers, reaction, enforceUnique, skipPush, currentUser!!)
             .precondition(plugins) { onSendReactionPrecondition(cid, currentUser, reaction) }
             .share(userScope) { SendReactionIdentifier(reaction, enforceUnique, cid) }
     }