+++ stream-chat-android-core/src/testFixtures/kotlin/io/getstream/chat/android/Mother.kt
@@ -237,6 +237,7 @@ public fun randomReaction(
     syncStatus: SyncStatus = randomSyncStatus(),
     extraData: MutableMap<String, Any> = mutableMapOf(),
     enforceUnique: Boolean = randomBoolean(),
+    emojiCode: String? = randomString(),
 ): Reaction = Reaction(
     messageId = messageId,
     type = type,
@@ -249,6 +250,7 @@ public fun randomReaction(
     syncStatus = syncStatus,
     extraData = extraData,
     enforceUnique = enforceUnique,
+    emojiCode = emojiCode,
 )
 
 public fun randomDraftMessageOrNull(): DraftMessage? = randomDraftMessage().takeIf { randomBoolean() }